+++ test/json_account.cpp
@@ -134,6 +134,8 @@ TEST(json_account, parser) {
         json::Action) override {
       FAIL();
     }
+    void operator()(server::Trace<json::IndexTickers> const &) override { FAIL(); }
+    void operator()(server::Trace<json::FundingRate> const &) override { FAIL(); }
     void operator()(server::Trace<json::Login> const &) override { FAIL(); }
     void operator()(server::Trace<json::Account> const &event) override {
       ++count_;