+++ test/json_balance_and_position.cpp
@@ -68,6 +68,8 @@ TEST(json_balance_and_position, parser) {
         json::Action) override {
       FAIL();
     }
+    void operator()(server::Trace<json::IndexTickers> const &) override { FAIL(); }
+    void operator()(server::Trace<json::FundingRate> const &) override { FAIL(); }
     void operator()(server::Trace<json::Login> const &) override { FAIL(); }
     void operator()(server::Trace<json::Account> const &) override { FAIL(); }
     void operator()(server::Trace<json::BalanceAndPosition> const &event) override {