+++ test/json_cancel_order_ack.cpp
@@ -56,6 +56,8 @@ TEST(json_cancel_order_ack, parser_success) {
         json::Action) override {
       FAIL();
     }
+    void operator()(server::Trace<json::IndexTickers> const &) override { FAIL(); }
+    void operator()(server::Trace<json::FundingRate> const &) override { FAIL(); }
     void operator()(server::Trace<json::Login> const &) override { FAIL(); }
     void operator()(server::Trace<json::Account> const &) override { FAIL(); }
     void operator()(server::Trace<json::BalanceAndPosition> const &) override { FAIL(); }