+++ test/json_funding_rate.cpp
@@ -22,20 +22,20 @@ auto create_trace_info() {
 }
 }  // namespace
 
-TEST(json_amend_order_ack, parser_success) {
+TEST(json_funding_rate, parser) {
   auto message = R"({)"
-                 R"("code":"0",)"
+                 R"("arg":{)"
+                 R"("channel":"funding-rate",)"
+                 R"("instId":"BTC-USD-SWAP")"
+                 R"(},)"
                  R"("data":[{)"
-                 R"("clOrdId":"CMAAF2IDAAAQAAFQDIHPD4Y3",)"
-                 R"("ordId":"393936310213439488",)"
-                 R"("reqId":"",)"
-                 R"("sCode":"0",)"
-                 R"("sMsg":"")"
+                 R"("fundingRate":"-0.00006384",)"
+                 R"("fundingTime":"1642665600000",)"
+                 R"("instId":"BTC-USD-SWAP",)"
+                 R"("instType":"SWAP",)"
+                 R"("nextFundingRate":"-0.00005")"
                  R"(})"
-                 R"(],)"
-                 R"("id":"2000002",)"
-                 R"("msg":"",)"
-                 R"("op":"amend-order")"
+                 R"(])"
                  R"(})";
   struct MyHandler final : public json::Parser::Handler {
     auto get_count() const { return count_; }
@@ -58,30 +58,25 @@ TEST(json_amend_order_ack, parser_success) {
       FAIL();
     }
     void operator()(server::Trace<json::IndexTickers> const &) override { FAIL(); }
-    void operator()(server::Trace<json::FundingRate> const &) override { FAIL(); }
+    void operator()(server::Trace<json::FundingRate> const &event) override {
+      ++count_;
+      auto &[trace_info, trades] = event;
+      auto &data = trades.data;
+      ASSERT_EQ(std::size(data), 1);
+      auto &d0 = data[0];
+      EXPECT_DOUBLE_EQ(d0.funding_rate, -0.00006384);
+      EXPECT_EQ(d0.funding_time, 1642665600000ms);
+      EXPECT_EQ(d0.inst_id, "BTC-USD-SWAP"sv);
+      EXPECT_EQ(d0.inst_type, json::InstrumentType::SWAP);
+      EXPECT_DOUBLE_EQ(d0.next_funding_rate, -0.00005);
+    }
     void operator()(server::Trace<json::Login> const &) override { FAIL(); }
     void operator()(server::Trace<json::Account> const &) override { FAIL(); }
     void operator()(server::Trace<json::BalanceAndPosition> const &) override { FAIL(); }
     void operator()(server::Trace<json::Positions> const &) override { FAIL(); }
     void operator()(server::Trace<json::Orders> const &) override { FAIL(); }
     void operator()(server::Trace<json::OrderAck> const &) override { FAIL(); }
-    void operator()(server::Trace<json::AmendOrderAck> const &event) override {
-      ++count_;
-      auto &[trace_info, amend_order_ack] = event;
-      EXPECT_EQ(amend_order_ack.code, 0);
-      auto &data = amend_order_ack.data;
-      ASSERT_EQ(std::size(data), 1);
-      auto &d0 = data[0];
-      EXPECT_EQ(d0.cl_ord_id, "CMAAF2IDAAAQAAFQDIHPD4Y3"sv);
-      EXPECT_EQ(d0.ord_id, "393936310213439488"sv);
-      EXPECT_EQ(d0.req_id, ""sv);
-      EXPECT_EQ(d0.s_code, 0);
-      EXPECT_EQ(d0.s_msg, ""sv);
-      EXPECT_EQ(d0.tag, ""sv);
-      EXPECT_EQ(amend_order_ack.id, "2000002"sv);
-      EXPECT_EQ(amend_order_ack.msg, ""sv);
-      EXPECT_EQ(amend_order_ack.op, json::Operation::AMEND_ORDER);
-    }
+    void operator()(server::Trace<json::AmendOrderAck> const &) override { FAIL(); }
     void operator()(server::Trace<json::CancelOrderAck> const &) override { FAIL(); }
 
    private: