+++ test/json_index_tickers.cpp
@@ -22,20 +22,23 @@ auto create_trace_info() {
 }
 }  // namespace
 
-TEST(json_amend_order_ack, parser_success) {
+TEST(json_index_tickers, parser) {
   auto message = R"({)"
-                 R"("code":"0",)"
+                 R"("arg":{)"
+                 R"("channel":"index-tickers",)"
+                 R"("instId":"BTC-USDT")"
+                 R"(},)"
                  R"("data":[{)"
-                 R"("clOrdId":"CMAAF2IDAAAQAAFQDIHPD4Y3",)"
-                 R"("ordId":"393936310213439488",)"
-                 R"("reqId":"",)"
-                 R"("sCode":"0",)"
-                 R"("sMsg":"")"
+                 R"("instId":"BTC-USDT",)"
+                 R"("idxPx":"41756.7",)"
+                 R"("open24h":"42427.3",)"
+                 R"("high24h":"42560.7",)"
+                 R"("low24h":"41147.5",)"
+                 R"("sodUtc0":"41663.1",)"
+                 R"("sodUtc8":"41984.2",)"
+                 R"("ts":"1642643951284")"
                  R"(})"
-                 R"(],)"
-                 R"("id":"2000002",)"
-                 R"("msg":"",)"
-                 R"("op":"amend-order")"
+                 R"(])"
                  R"(})";
   struct MyHandler final : public json::Parser::Handler {
     auto get_count() const { return count_; }
@@ -57,7 +60,21 @@ TEST(json_amend_order_ack, parser_success) {
         json::Action) override {
       FAIL();
     }
-    void operator()(server::Trace<json::IndexTickers> const &) override { FAIL(); }
+    void operator()(server::Trace<json::IndexTickers> const &event) override {
+      ++count_;
+      auto &[trace_info, trades] = event;
+      auto &data = trades.data;
+      ASSERT_EQ(std::size(data), 1);
+      auto &d0 = data[0];
+      EXPECT_EQ(d0.inst_id, "BTC-USDT"sv);
+      EXPECT_DOUBLE_EQ(d0.idx_px, 41756.7);
+      EXPECT_DOUBLE_EQ(d0.open24h, 42427.3);
+      EXPECT_DOUBLE_EQ(d0.high24h, 42560.7);
+      EXPECT_DOUBLE_EQ(d0.low24h, 41147.5);
+      EXPECT_DOUBLE_EQ(d0.sod_utc0, 41663.1);
+      EXPECT_DOUBLE_EQ(d0.sod_utc8, 41984.2);
+      EXPECT_EQ(d0.ts, 1642643951284ms);
+    }
     void operator()(server::Trace<json::FundingRate> const &) override { FAIL(); }
     void operator()(server::Trace<json::Login> const &) override { FAIL(); }
     void operator()(server::Trace<json::Account> const &) override { FAIL(); }
@@ -65,23 +82,7 @@ TEST(json_amend_order_ack, parser_success) {
     void operator()(server::Trace<json::Positions> const &) override { FAIL(); }
     void operator()(server::Trace<json::Orders> const &) override { FAIL(); }
     void operator()(server::Trace<json::OrderAck> const &) override { FAIL(); }
-    void operator()(server::Trace<json::AmendOrderAck> const &event) override {
-      ++count_;
-      auto &[trace_info, amend_order_ack] = event;
-      EXPECT_EQ(amend_order_ack.code, 0);
-      auto &data = amend_order_ack.data;
-      ASSERT_EQ(std::size(data), 1);
-      auto &d0 = data[0];
-      EXPECT_EQ(d0.cl_ord_id, "CMAAF2IDAAAQAAFQDIHPD4Y3"sv);
-      EXPECT_EQ(d0.ord_id, "393936310213439488"sv);
-      EXPECT_EQ(d0.req_id, ""sv);
-      EXPECT_EQ(d0.s_code, 0);
-      EXPECT_EQ(d0.s_msg, ""sv);
-      EXPECT_EQ(d0.tag, ""sv);
-      EXPECT_EQ(amend_order_ack.id, "2000002"sv);
-      EXPECT_EQ(amend_order_ack.msg, ""sv);
-      EXPECT_EQ(amend_order_ack.op, json::Operation::AMEND_ORDER);
-    }
+    void operator()(server::Trace<json::AmendOrderAck> const &) override { FAIL(); }
     void operator()(server::Trace<json::CancelOrderAck> const &) override { FAIL(); }
 
    private: