+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+# Added
+
+* Add funding rate and index price (#161)
+
 ## 0.8.1 &ndash; 2022-01-16
 
 ## 0.8.0 &ndash; 2022-01-12